backend/requirements.txt
flask==3.0.0
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
tensorflow==2.14.0
transformers==4.35.0
pandas==2.1.2
numpy==1.26.1